rename: StudyMember studyEntity로 이동
