+++ src/main/java/com/grepp/spring/app/model/member/repository/StudyMemberRepositoryCustomImpl.java
@@ -6,7 +6,7 @@ import com.grepp.spring.app.model.auth.code.Role;
 import com.grepp.spring.app.model.member.code.StudyRole;
 import com.grepp.spring.app.model.member.entity.QMember;
 import com.grepp.spring.app.model.member.entity.QStudyMember;
-import com.grepp.spring.app.model.member.entity.StudyMember;
+import com.grepp.spring.app.model.study.entity.StudyMember;
 import com.grepp.spring.app.model.study.entity.QStudy;
 import com.grepp.spring.app.model.study.entity.QStudySchedule;
 import com.querydsl.jpa.impl.JPAQueryFactory;
