+++ src/main/java/com/grepp/spring/app/model/study/entity/Attendance.java
@@ -1,6 +1,5 @@
 package com.grepp.spring.app.model.study.entity;
 
-import com.grepp.spring.app.model.member.entity.StudyMember;
 import jakarta.persistence.*;
 import lombok.*;
