+++ src/main/java/com/grepp/spring/app/model/study/entity/StudyMember.java
@@ -1,7 +1,7 @@
-package com.grepp.spring.app.model.member.entity;
+package com.grepp.spring.app.model.study.entity;
 
 import com.grepp.spring.app.model.member.code.StudyRole;
-import com.grepp.spring.app.model.study.entity.Study;
+import com.grepp.spring.app.model.member.entity.Member;
 import com.grepp.spring.infra.entity.BaseEntity;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
