+++ src/test/java/com/grepp/spring/app/model/quiz/service/SurvivalResultServiceTest.java
@@ -2,7 +2,7 @@ package com.grepp.spring.app.model.quiz.service;
 
 import com.grepp.spring.app.controller.api.quiz.payload.SurvivalResultRequest;
 import com.grepp.spring.app.model.member.entity.Member;
-import com.grepp.spring.app.model.member.entity.StudyMember;
+import com.grepp.spring.app.model.study.entity.StudyMember;
 import com.grepp.spring.app.model.member.repository.MemberRepository;
 import com.grepp.spring.app.model.member.repository.StudyMemberRepository;
 import com.grepp.spring.app.model.quiz.entity.QuizRecord;
